+++ CHANGELOG.md
@@ -2,7 +2,9 @@ This is a user-focused changelog. The entries here are things that affect gamepl
 
 Entries in bold are ones important enough that I would have put them in [the old changelog](./CHANGELOG-old.md).
 
-* DEV: Added eFG% to stat tables in BBGM. Thanks @domini7! [[1]](https://github.com/dumbmatter/gm-games/pull/310)
+* 2020-09-07: New challenge mode "Lose best player every season", where at the end of the playoffs every season, the best player on your team will either retire (if he's a real player) or die a tragic death (if he's a random player). Thanks @luso97! [[1]](https://github.com/dumbmatter/gm-games/pull/311/files) [[2]]()
+
+* 2020-09-07: Added eFG% to stat tables in BBGM. Thanks @domini7! [[1]](https://github.com/dumbmatter/gm-games/pull/310)
 
 * 2020-09-03: Fixed team/player stat distributions pages for FBGM, and added additional stats in BBGM. Thanks @luso97! [[1]](https://github.com/dumbmatter/gm-games/pull/306)